migrate bvhcast
